+++ src/rename/mod.rs
@@ -539,7 +539,10 @@ impl Backend {
             SymbolKind::ConstantReference { name } => Some((name.clone(), range)),
             SymbolKind::NamespaceDeclaration { name } => Some((name.clone(), range)),
             SymbolKind::SelfStaticParent { .. } => None,
-            SymbolKind::LaravelStringKey { .. } => None,
+            SymbolKind::LaravelStringKey { .. }
+            | SymbolKind::Keyword
+            | SymbolKind::CastType
+            | SymbolKind::Comment => None,
         }
     }